requirements.txt
langchain-chroma
langchain-huggingface
langchain-openai
python-dotenv
sentence-transformers
tqdm
httpcore
langchain-community
langchain-core
unstructured
pandas
openpyxl
python-pptx
python-docx
nltk
